LPS-115016 Unify ArtifactListeners and rename
